+++ src/app/(sub)/join/components/JoinContentTextMessage.client.tsx
