+++ src/app/(sub)/join/components/JoinHeader.client.tsx
