+++ src/app/(sub)/join/components/JoinTitleTextMessage.client.tsx
