+++ src/app/(sub)/join/funnels/step1/Step1Component.client.tsx
