+++ src/app/(sub)/join/funnels/step1/components/NoticeSection.client.tsx
