+++ src/app/(sub)/join/funnels/step1/components/NumberForm.client.tsx
@@ -3,7 +3,7 @@
 import { useJoinContext } from '../../../components/JoinContext.client';
 import { formatNumber, formatNumberBackSpace } from '../util';
 import { useSMSMutation } from '@/apis/auth';
-import { type SignUpState } from '@/app/join/type';
+import { type SignUpState } from '@/app/(sub)/join/type';
 import { Button, ButtonGroup } from '@/components/Button';
 import { Spacing } from '@/components/common/Spacing';
 import { TextFieldController } from '@/components/TextField';
