+++ src/app/(sub)/join/funnels/step1/components/NumberVerifyForm.client.tsx
@@ -10,7 +10,7 @@ import { setTokenAtCookie } from '@/utils/auth/tokenController';
 import { useRouter } from 'next/navigation';
 import { type SubmitHandler } from 'react-hook-form';
 
-import type { SignUpState } from '@/app/join/type';
+import type { SignUpState } from '@/app/(sub)/join/type';
 
 export default function NumberVerifyForm() {
   const router = useRouter();
