+++ src/app/(sub)/join/funnels/step1/type.ts
