+++ src/app/(sub)/join/funnels/step2/Step2Component.client.tsx
