+++ src/app/(sub)/join/funnels/step2/components/SchoolSection.client.tsx
