+++ src/app/(sub)/join/funnels/step2/components/SearchResultSection.client.tsx
