+++ src/app/(sub)/join/funnels/step2/type.ts
