+++ src/app/(sub)/join/funnels/step3/components/EmailForm.client.tsx
@@ -3,8 +3,8 @@ import EmailSection from './EmailSection.client';
 import { useTimerContext } from './TimerContext.client';
 import { useFunnelContext } from '../../JoinFunnel';
 import { useEmailMutation } from '@/apis/auth';
-import { useJoinContext } from '@/app/join/components/JoinContext.client';
-import { SignUpState } from '@/app/join/type';
+import { useJoinContext } from '@/app/(sub)/join/components/JoinContext.client';
+import { SignUpState } from '@/app/(sub)/join/type';
 import { Button, ButtonGroup } from '@/components/Button';
 import BottomFixedDiv from '@/components/common/BottomFixedDiv';
 import { Spacing } from '@/components/common/Spacing';
