+++ src/app/(sub)/join/funnels/step3/components/NoticeSection.client.tsx
