+++ src/app/(sub)/join/funnels/step4/components/ImageSection.client.tsx
