+++ src/app/(sub)/join/funnels/step4/components/InputForm.client.tsx
@@ -4,7 +4,7 @@ import GenderSection from './GenderSection.client';
 import ImageSection from './ImageSection.client';
 import NicknameSection from './NicknameSection.client';
 import { useFunnelContext } from '../../JoinFunnel';
-import { useJoinContext } from '@/app/join/components/JoinContext.client';
+import { useJoinContext } from '@/app/(sub)/join/components/JoinContext.client';
 import { BottomFixedButton } from '@/components/common/Button';
 
 export default function InputForm() {
