+++ src/app/(sub)/join/funnels/step4/components/NicknameSection.client.tsx
