+++ src/app/(sub)/join/funnels/step5/components/InputForm.client.tsx
@@ -3,13 +3,13 @@
 import PersonalitySection from './PersonalitySection.client';
 import { formatDate } from '../util';
 import { useSignUpMutation } from '@/apis/auth';
-import { useJoinContext } from '@/app/join/components/JoinContext.client';
+import { useJoinContext } from '@/app/(sub)/join/components/JoinContext.client';
 import { BottomFixedButton } from '@/components/common/Button';
 import { personalityList } from '@/constants/personalityList';
 import { setTokenAtCookie } from '@/utils/auth/tokenController';
 import { useRouter } from 'next/navigation';
 
-import type { SignUpState } from '@/app/join/type';
+import type { SignUpState } from '@/app/(sub)/join/type';
 import type { GenderType } from '@/types';
 
 export default function InputForm() {
