+++ src/app/(sub)/join/funnels/step5/components/PersonalityItem.client.tsx
@@ -1,5 +1,5 @@
 'use client';
-import { useJoinContext } from '@/app/join/components/JoinContext.client';
+import { useJoinContext } from '@/app/(sub)/join/components/JoinContext.client';
 import clsx from 'clsx';
 import { memo } from 'react';
