+++ src/app/(sub)/onboarding/components/OnboardingSpeechBubble.tsx
