--- src/app/board/layout.tsx
@@ -1,8 +1,0 @@
-export default function BoardLayout({ children }: { children: React.ReactNode }) {
-  return (
-    <div>
-      <div className="relative h-full bg-white px-24">{children}</div>
-      {/* BottomNavigationBar */}
-    </div>
-  );
-}
