--- src/app/board/page.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Board = () => {
-  return <div></div>;
-};
-
-export default Board;
